refactor(wallet): extract store conformance shared examples for downstream gems (#591)
Move the storage adapter shared examples and wallet-level operation specs
into `lib/bsv/wallet/testing/` so downstream adapter gems can load them
via `require 'bsv/wallet/testing/store_conformance'` instead of using a
fragile relative path into bsv-wallet's spec tree.
